@@ -1,70 +1,43 @@
-name: trem
-description: trem
+name: trem_pocket
+description: A new Flutter project.
 
 publish_to: 'none'
+
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.2 <3.0.0"
+  sdk: '>=2.18.6 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_background_service: ^2.1.3
-  device_info_plus: ^4.1.2
-  web_socket_channel: ^2.2.0
-  audioplayers: ^1.0.1
-  intl: ^0.17.0
-  perfect_volume_control: ^1.0.5
-  uuid: ^3.0.6
-  hive_flutter: ^1.1.0
-  hive: ^2.2.3
-  firebase_messaging: ^12.0.3
-  firebase_core: ^1.21.0
-  loggy: ^2.0.1+1
-  flutter_loggy: ^2.0.1
-  http: ^0.13.5
-  geolocator: ^9.0.1
-  clipboard: ^0.1.3
-  flutter_map: ^2.2.0
-  latlong2: ^0.8.1
-
   cupertino_icons: ^1.0.5
+  http: ^0.13.5
+  flutter_map: ^3.1.0
+  latlong2: ^0.8.1
+  firebase_core: ^2.4.0
+  firebase_messaging: ^14.2.0
+  clipboard: ^0.1.3
+  flutter_background_service: ^2.4.6
+  flutter_local_notifications: ^13.0.0
+  device_info_plus: ^8.0.0
+  shared_preferences: ^2.0.15
+  package_info_plus: ^3.0.2
+  markdown_widget: ^1.3.0+2
+  url_launcher: ^6.1.7
+  flutter_switch: ^0.3.2
+  store_checker: ^1.1.0
+  alarmplayer: ^1.0.5
+  auto_start_flutter: ^0.1.1
+  android_autostart: ^0.0.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
   flutter_lints: ^2.0.1
 
 flutter:
   uses-material-design: true
   assets:
-    - audios/
-    - audios/1/
-
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware.
-
-  # For details regarding adding assets from package dependencies, see
-  # https://flutter.dev/assets-and-images/#from-packages
-
-  # To add custom fonts to your application, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts from package dependencies,
-  # see https://flutter.dev/custom-fonts/#from-packages
+    - assets/images/
+    - assets/audios/
